refactor: 아이 정보 컴포넌트 view로 분리
