+++ src/components/MyPetsTemplate/NoPets.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import PetRegisterButton from 'components/MyPets/PetRegisterButton';
+import PetRegisterButton from 'components/MyPetsTemplate/PetRegisterButton';
 import { INFO_MESSAGES, USER_ACTIONS } from './constants';
 
 function NoPets() {
