+++ src/components/MyPetsTemplate/index.jsx
@@ -1,19 +1,13 @@
-/* eslint-disable no-shadow */
-import React, { Suspense } from 'react';
+import React from 'react';
 
-import Spinner from 'components/Spinner';
-import PetRegisterButton from 'components/MyPets/PetRegisterButton';
-
+import PetRegisterButton from 'components/MyPetsTemplate/PetRegisterButton';
+import PetList from './PetList';
 import plus from '../../assets/plus.svg';
 
-const FetchPets = React.lazy(() => import('./FetchPets'));
-
-function MyPets() {
+function MyPetsTemplate() {
   return (
     <div className="h-full">
-      <Suspense fallback={<Spinner />}>
-        <FetchPets />
-      </Suspense>
+      <PetList />
       <PetRegisterButton className="flex items-center justify-center gap-x-2 mt-1 py-5 bg-white border border-dashed border-orange-400 rounded-2xl">
         <img src={plus} alt="add" />
         <span className="text-solo-label-pc text-orange-400 font-semibold">
@@ -24,4 +18,4 @@ function MyPets() {
   );
 }
 
-export default MyPets;
+export default MyPetsTemplate;
