+++ src/view/MyPets.tsx
@@ -1,0 +1,14 @@
+import React, { Suspense } from 'react';
+import Spinner from 'components/Spinner';
+
+const MyPetsTemplate = React.lazy(() => import('components/MyPetsTemplate'));
+
+function MyPets() {
+  return (
+    <Suspense fallback={<Spinner />}>
+      <MyPetsTemplate />
+    </Suspense>
+  );
+}
+
+export default MyPets;
